Single vs. Multi Play
Can now toggle between playing a singular sound effect and playing multiple effects at the same time. However, playing multiple sounds exploits a behavior that I believe causes a memory leak. For example, playing the temp file through the main player rather than the playback player causes it to be locked up in the main player and not able to be opened anymore. Need to fix this.